Added basic functionality that answers challenge questions one through three.
A dropdown box is used to allow the user to select which function they wish to perform. New code added to the bottom of index.php to retrieve and validate the user's inputs. The code calls the date functions defined in date_functions.php. Unit tests will need to be added to test these date functions upon completion of the five questions.