Added the ability to specify timezones to all three date functions.
Timezones can be set either before passing a TimeDate object to the function or as the 4th and 5th parameters. If a timezone is not specified then the function will use UTC-0. Helper functions have been added in order to further validate data passed to the function as well as to combine separately created DateTime objects with their respective timezones.

num_week_days and num_weeks have been modified to work with times as previously they would not take the time into account. Now if the time is less than 24 hours, it is no longer considered a day in the function.